Fix an assumed typo and clean up the repo
I hope this is OK, but I went ahead and assumed that you wanted the
whole project to be GPLv3 and changed it. I also deleted all the empty
files.

I went ahead and added a simple .gitattributes and added a few entries
into the .gitignore file for Mac and Windows users ( I personally am
using a Cygwin setup for quick testing). If you have any concerns or
questions let me know. Hope this helps.