See description
I didn't actually change that much, but because GitHub views a change of
encoding or EOL as a change, most of the files appear to be changed.

Anyway, all I did was update the .gitignore file to be way more
universal, and I also updated all the copyright messages to include
GPLv3 and you as the creator.

The only other things I changed were cosmetic changes to blank
documents. I filled out the changelog, author, and other files to be
more complete.

Everything else should be largely unchanged, except I added a newline at
the end of every document so that certain applications and Linux
distributions would be more happy.

Hope everything goes OK with this commit. :smiley: